fix(typescript): 将 store 目录重命名为 stores 并更新导入
为了保持一致性，将 'store' 目录重命名为 'stores'。更新了所有相关的导入路径和 Vite 别名配置。添加了一个 Husky commit-msg 钩子，用于提交消息 linting。
