+++ .husky/commit-msg
@@ -1,0 +1,2 @@
+#!/usr/bin/env sh
+npx --no-install commitlint --edit "$1"
