+++ src/pages/HomePage.vue
@@ -7,13 +7,12 @@
 
 <script lang="ts" setup>
 import { ref } from 'vue';
-import { useUserStore } from '@/store/modules/user';
+import { useUserStore } from '@/stores/modules/user';
 // import type { UserType } from '@/types/user';
 
 const userStore = useUserStore();
 
 const userInfo = ref(userStore.userData);
-
 </script>
 
 <style lang="scss" scoped></style>
